requirements.txt
certifi==2023.11.17; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
idna==3.6; python_version >= '3.5'
python-dotenv==1.0.0; python_version >= '3.8'
requests==2.31.0; python_version >= '3.7'
urllib3==2.1.0; python_version >= '3.8'